+++ api-base/detectOther.ts
@@ -1,4 +1,4 @@
-// Copyright 2017-2022 @polkadot/api-base authors & contributors
+// Copyright 2017-2023 @polkadot/api-base authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 import { packageInfo as rpcInfo } from 'https://deno.land/x/polkadot/rpc-core/packageInfo.ts';