+++ api-contract/detectOther.ts
@@ -1,4 +1,4 @@
-// Copyright 2017-2022 @polkadot/api-contract authors & contributors
+// Copyright 2017-2023 @polkadot/api-contract authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 import { packageInfo as apiInfo } from 'https://deno.land/x/polkadot/api/packageInfo.ts';