+++ rpc-provider/mock/mockWs.ts
@@ -1,4 +1,4 @@
-// Copyright 2017-2022 @polkadot/rpc-provider authors & contributors
+// Copyright 2017-2023 @polkadot/rpc-provider authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 import { Server, WebSocket } from 'https://esm.sh/mock-socket@9.1.5';