+++ types/metadata/util/getSiName.ts
@@ -1,4 +1,4 @@
-// Copyright 2017-2022 @polkadot/types authors & contributors
+// Copyright 2017-2023 @polkadot/types authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 import type { ILookup } from 'https://deno.land/x/polkadot/types-create/types/index.ts';